+++ package.json
@@ -24,6 +24,7 @@
     "@vercel/og": "^0.5.8",
     "autoprefixer": "10.4.14",
     "class-variance-authority": "^0.6.1",
+    "classnames": "^2.3.2",
     "clsx": "^1.2.1",
     "contentlayer": "^0.3.3",
     "date-fns": "^2.30.0",
